requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-computed-property==0.3.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
xlrd==2.0.1
